Continued moving non-implementation specific details of testing into enginetests.go
